requirements.txt
alembic==1.0.10
Click==7.0
decorator==4.4.0
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Images==3.0.1
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
PIL==1.1.7
Pillow==6.0.0
pillowcase==2.0.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
SQLAlchemy-Utils==0.33.11
validators==0.12.6
virtualenv==16.5.0
Werkzeug==0.15.2
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
